refactor: Fix types and imports
