+++ package.json
@@ -127,5 +127,10 @@
   },
   "nyc": {
     "report-dir": "coverage/cypress"
+  },
+  "pnpm": {
+    "patchedDependencies": {
+      "cytoscape@3.28.1": "patches/cytoscape@3.28.1.patch"
+    }
   }
 }
