+++ packages/mermaid-example-diagram/package.json
@@ -39,15 +39,10 @@
   },
   "dependencies": {
     "@braintree/sanitize-url": "^6.0.1",
-    "cytoscape": "^3.23.0",
-    "cytoscape-cose-bilkent": "^4.1.0",
-    "cytoscape-fcose": "^2.1.0",
     "d3": "^7.0.0",
-    "khroma": "^2.0.0",
-    "non-layered-tidy-tree-layout": "^2.0.2"
+    "khroma": "^2.0.0"
   },
   "devDependencies": {
-    "@types/cytoscape": "^3.19.9",
     "concurrently": "^8.0.0",
     "rimraf": "^5.0.0",
     "mermaid": "workspace:*"
